change arch, make one core interface
