+++ include/mppic/optimizers/xtensor/critics/goal_critic.hpp
@@ -1,18 +1,23 @@
 // Copyright 2022 FastSense, Samsung Research
 #pragma once
 
-
-#include "mppic/critic_function.hpp"
+#include "mppic/optimizers/xtensor/critic_function.hpp"
+#include "mppic/optimizers/xtensor/models/state.hpp"
 #include "mppic/utils.hpp"
 
 namespace mppi::critics
 {
 
-class PreferForwardCritic : public CriticFunction
+class GoalCritic : public CriticFunction
 {
 public:
   void initialize() override;
 
+  /**
+   * @brief Evaluate cost related to goal following
+   *
+   * @param costs [out] add reference cost values to this tensor
+   */
   void evalScore(models::CriticFunctionData & data) override;
 
 protected:
