+++ include/mppic/optimizers/xtensor/critics/reference_trajectory_critic.hpp
@@ -1,8 +1,8 @@
 // Copyright 2022 FastSense, Samsung Research
 #pragma once
 
-#include "mppic/critic_function.hpp"
-#include "mppic/models/state.hpp"
+#include "mppic/optimizers/xtensor/critic_function.hpp"
+#include "mppic/optimizers/xtensor/models/state.hpp"
 #include "mppic/utils.hpp"
 
 namespace mppi::critics
