+++ include/mppic/optimizers/xtensor/critics/twirling_critic.hpp
@@ -1,7 +1,7 @@
 // Copyright 2022 FastSense, Samsung Research
 #pragma once
 
-#include "mppic/critic_function.hpp"
+#include "mppic/optimizers/xtensor/critic_function.hpp"
 #include "mppic/utils.hpp"
 
 namespace mppi::critics
