+++ include/mppic/utils.hpp
@@ -11,15 +11,19 @@
 #include <xtensor/xmath.hpp>
 #include <xtensor/xview.hpp>
 
-#include "geometry_msgs/msg/twist_stamped.hpp"
-#include "mppic/models/control_sequence.hpp"
-#include "nav2_util/node_utils.hpp"
-#include "nav2_core/goal_checker.hpp"
-#include "nav_msgs/msg/path.hpp"
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_lifecycle/lifecycle_node.hpp"
 #include "tf2/utils.h"
 #include "tf2_geometry_msgs/tf2_geometry_msgs.hpp"
+
+#include "geometry_msgs/msg/twist_stamped.hpp"
+#include "nav_msgs/msg/path.hpp"
+
+#include "rclcpp/rclcpp.hpp"
+#include "rclcpp_lifecycle/lifecycle_node.hpp"
+
+#include "nav2_util/node_utils.hpp"
+#include "nav2_core/goal_checker.hpp"
+
+#include "mppic/optimizers/xtensor/models/control_sequence.hpp"
 
 namespace mppi::utils
 {
