+++ src/optimizers/xtensor/critics/goal_angle_critic.cpp
@@ -1,5 +1,5 @@
 // Copyright 2022 FastSense, Samsung Research
-#include "mppic/critics/goal_angle_critic.hpp"
+#include "mppic/optimizers/xtensor/critics/goal_angle_critic.hpp"
 
 namespace mppi::critics
 {
